daemon: move tracemalloc `jsonrpc_` methods to an individual module
