+++ lbry/extras/daemon/daemon.py
@@ -234,6 +234,7 @@ from lbry.extras.daemon.daemon_txo import Daemon_txo
 from lbry.extras.daemon.daemon_utxo import Daemon_utxo
 from lbry.extras.daemon.daemon_blob import Daemon_blob
 from lbry.extras.daemon.daemon_peer import Daemon_peer
+from lbry.extras.daemon.daemon_tracemalloc import Daemon_tracemalloc
 
 HISTOGRAM_BUCKETS = (
     .005, .01, .025, .05, .075, .1, .25, .5, .75, 1.0, 2.5, 5.0, 7.5, 10.0, 15.0, 20.0, 30.0, 60.0, float('inf')
@@ -245,7 +246,7 @@ class Daemon(Daemon_base, Daemon_get, Daemon_settings, Daemon_wallet,
              Daemon_purchase, Daemon_claim, Daemon_channel,
              Daemon_stream, Daemon_collection, Daemon_support,
              Daemon_transaction, Daemon_txo, Daemon_utxo, Daemon_blob,
-             Daemon_peer):
+             Daemon_peer, Daemon_tracemalloc):
     """
     LBRYnet daemon, a jsonrpc interface to lbry functions
     """
@@ -908,81 +909,9 @@ class Daemon(Daemon_base, Daemon_get, Daemon_settings, Daemon_wallet,
     TRACEMALLOC_DOC = """
     Controls and queries tracemalloc memory tracing tools for troubleshooting.
     """
-
-    def jsonrpc_tracemalloc_enable(self):  # pylint: disable=no-self-use
-        """
-        Enable tracemalloc memory tracing
-
-        Usage:
-            jsonrpc_tracemalloc_enable
-
-        Options:
-            None
-
-        Returns:
-            (bool) is it tracing?
-        """
-        tracemalloc.start()
-        return tracemalloc.is_tracing()
-
-    def jsonrpc_tracemalloc_disable(self):  # pylint: disable=no-self-use
-        """
-        Disable tracemalloc memory tracing
-
-        Usage:
-            jsonrpc_tracemalloc_disable
-
-        Options:
-            None
-
-        Returns:
-            (bool) is it tracing?
-        """
-        tracemalloc.stop()
-        return tracemalloc.is_tracing()
-
-    def jsonrpc_tracemalloc_top(self, items: int = 10):  # pylint: disable=no-self-use
-        """
-        Show most common objects, the place that created them and their size.
-
-        Usage:
-            jsonrpc_tracemalloc_top [(<items> | --items=<items>)]
-
-        Options:
-            --items=<items>               : (int) maximum items to return, from the most common
-
-        Returns:
-            (dict) dictionary containing most common objects in memory
-            {
-                "line": (str) filename and line number where it was created,
-                "code": (str) code that created it,
-                "size": (int) size in bytes, for each "memory block",
-                "count" (int) number of memory blocks
-            }
-        """
-        if not tracemalloc.is_tracing():
-            raise Exception("Enable tracemalloc first! See 'tracemalloc set' command.")
-        stats = tracemalloc.take_snapshot().filter_traces((
-            tracemalloc.Filter(False, "<frozen importlib._bootstrap>"),
-            tracemalloc.Filter(False, "<unknown>"),
-            # tracemalloc and linecache here use some memory, but thats not relevant
-            tracemalloc.Filter(False, tracemalloc.__file__),
-            tracemalloc.Filter(False, linecache.__file__),
-        )).statistics('lineno', True)
-        results = []
-        for stat in stats:
-            frame = stat.traceback[0]
-            filename = os.sep.join(frame.filename.split(os.sep)[-2:])
-            line = linecache.getline(frame.filename, frame.lineno).strip()
-            results.append({
-                "line": f"{filename}:{frame.lineno}",
-                "code": line,
-                "size": stat.size,
-                "count": stat.count
-            })
-            if len(results) == items:
-                break
-        return results
+    # jsonrpc_tracemalloc_enable
+    # jsonrpc_tracemalloc_disable
+    # jsonrpc_tracemalloc_top
 
     COMMENT_DOC = """
     View, create and abandon comments.
